auto-analyst-backend/requirements.txt
aiofiles==24.1.0
beautifulsoup4==4.13.4
chardet==5.2.0
dspy==3.0.3
litellm==1.75.2
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.7
FastAPI-SQLAlchemy==0.2.1
fastapi-sso==0.16.0
groq==0.18.0
gunicorn==23.0.0
huggingface-hub==0.30.2
joblib==1.4.2
llama-cloud==0.1.19
llama-cloud-services==0.6.21
llama-index==0.12.14
llama-index-agent-openai==0.4.2
llama-index-cli==0.4.1
llama-index-core==0.12.34.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.14
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
Markdown==3.7
matplotlib==3.10.0
matplotlib-inline==0.1.7
numpy==2.2.2
openpyxl==3.1.2
xlrd==2.0.1
openai==1.97.0
pandas==2.2.3
polars==1.31.0
pillow==11.1.0
plotly==5.24.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
SQLAlchemy==2.0.37
statsmodels==0.14.4
tabulate==0.9.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
urllib3==2.4.0
uvicorn==0.29.0
websockets>=13.1.0
wheel==0.45.1
xgboost-cpu==3.0.2
bokeh==3.7.3
pymc==5.23.0
lightgbm==4.6.0
arviz==0.21.0
optuna==4.3.0
litellm[proxy]
duckdb==1.3.2